requirements.txt
# Core dependencies for XAI Scoring Framework
streamlit==1.28.1
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
plotly==5.17.0
matplotlib==3.8.2
openpyxl==3.1.2
python-docx==1.1.0
openai==1.3.7

# Web framework
flask==2.3.3
werkzeug==2.3.7
waitress==3.0.0

# API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Development and testing
pytest==7.4.3 